chore: rename file and add documentation
